Updated routes and removed pageRoutes from server.js into the routes/index.js folder
