+++ routes/pageRoutes.js
@@ -1,0 +1,12 @@
+const router = require('express').Router();
+const path= require("path")
+
+router.get('/', (req,res) => {
+    res.sendFile(path.join(__dirname, '../public/index.html'));
+});
+
+router.get('/notes', (req,res) => {
+    res.sendFile(path.join(__dirname, '../public/notes.html'));
+});
+
+module.exports = router
