+++ server.js
@@ -1,24 +1,15 @@
 const express = require('express');
-const path = require('path');
-const api = require('./routes/api/index');
-const fs = require('fs');
+const routes = require('./routes');
 
 var PORT = process.env.PORT || 3001;
 const app = express();
 
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
-app.use('/api', api);
 
 app.use(express.static('public'));
 
-app.get('/', (req,res) => 
-res.sendFile(path.join(__dirname, '/public/index.html'))
-);
-
-app.get('/notes', (req, res) =>
-res.sendFile(path.join(__dirname, 'public/notes.html'))
-);
+app.use(routes)
 
 app.listen(PORT, () =>
 console.log(`App listening at http://localhost:${PORT}`)
